split cob-autosens into two separate files
