+++ bin/oref0-detect-sensitivity.js
@@ -17,7 +17,7 @@
 
 var basal = require('oref0/lib/profile/basal');
 var get_iob = require('oref0/lib/iob');
-var detect = require('oref0/lib/determine-basal/cob-autosens');
+var detect = require('oref0/lib/determine-basal/autosens');
 
 if (!module.parent) {
     var detectsensitivity = init();
