+++ lib/determine-basal/autosens.js
