+++ lib/meal/total.js
@@ -1,5 +1,5 @@
 var tz = require('moment-timezone');
-var calcMealCOB = require('oref0/lib/determine-basal/cob-autosens');
+var calcMealCOB = require('oref0/lib/determine-basal/cob');
 
 function recentCarbs(opts, time) {
     var treatments = opts.treatments;
